third_party/blink/web_tests/platform/generic/fast/xmlhttprequest/xmlhttprequest-responseXML-xml-document-responsetype-expected.txt

This tests the XMLHttpRequest responseXML loading an XML document with responseType "document".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.responseXML is non-null.
PASS xhr.responseXML instanceof Document is true
PASS id3 is non-null.
PASS id3.textContent is "Three"
PASS successfullyParsed is true

TEST COMPLETE
